deps: remove whatwg-url override
This commit removes the final override from the package.json file.
The whatwg-url override was added to silence Node core deprecation
warning DEP0040. While this silences the warning in the CI, it does
not help users of this module.

I have also opened https://github.com/nodejs/node/pull/56632. If
landed, that will make the warning go away for users of this
library.