docs: correct directory and set entryPointStrategy
The directory where to generate the docs from was wrong, I'm not sure
when this changed but probably during the move from release-0.x to
release-1.x.

Also the `entryPointStrategy` needed to be set so that we are able to
include these files to generate the docs even if our ts-config doesn't
include them.
This was the default behavior a couple of releases of typedoc ago but
was changed.